Build fix.

+++ src/common/Lgi/FontSelect.cpp
@@ -323,7 +323,7 @@ bool LFontSelect::Serialize(void *Data, int DataLen, bool Write)
 		}
 		if (T[1])
 		{
-			Size = T[1].Int();
+			Size = (int)T[1].Int();
 		}
 
 		Bold = Underline = Italic = false;
